+++ crates/voicevox_core_python_api/python/voicevox_core/blocking.py
@@ -1,8 +1,5 @@
 """ブロッキング版API。"""
 
-# TODO: `AudioFeature`を復活させる
-# https://github.com/VOICEVOX/voicevox_core/issues/970
-
 # pyright: reportMissingModuleSource=false
 from ._rust.blocking import (
     Onnxruntime,
